As to library configuration, a new directory has been added within test dependencies, refering to "Data". In this case, it is going to store all of the resources required to properly server the target webpage.
In order to handle those data dependencies, some changes have been done on Common shell files library, but those changes have not been commited yet, so DEBUG version is being used by now.

Lastly, a the new dependency has been added within the test dependencies section. CSF does not require those to include a URL nor a version (it can make sense in cases as the current one, but it won't if data was a text, for example), so JMS library compliance is not required explicitly (would you imagine having to write a config.xml file in order to work with a text file?).